% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_opinions.R
\name{dfm_opinions}
\alias{dfm_opinions}
\title{Create a dfm with opinion features}
\usage{
dfm_opinions(x, dictionary, dictionary_sentiment, negation = negations(),
  ...)
}
\arguments{
\item{x}{a tokens object}

\item{dictionary}{a dictionary obejct to specify topics or entities the
opinions are about.}

\item{dictionary_sentiment}{a dictionary object for sentiment analysis where
first and second keys are negative and positive sentiment, respectively.}

\item{negation}{tokens indicating negations in sentences}

\item{...}{additional argumets passed to \code{tokens_kwic}}
}
\description{
Create a dfm with opinion features
}
