% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_kwic.R
\name{tokens_kwic}
\alias{tokens_kwic}
\title{Segment documents into windows}
\usage{
tokens_kwic(x, pattern, window = 5, valuetype = c("glob", "regex",
  "fixed"), case_insensitive = TRUE, use_docvars = TRUE, ...)
}
\description{
Segment documents into windows
}
\examples{
require(quanteda)
toks <- tokens(data_corpus_irishbudget2010)
toks_kwic <- tokens_kwic(toks, "tax")
}
